Change the way that RNG streams are determined.  The underlying (#72)
RNG library uses 2x64bit numbers for the key and 2x64bit numbers
for the counter.  These are now set according to:

key1 = realization * 2^32 + telescope * 2^16 + component
key2 = obsindx * 2^32 + detindx
counter1 = currently unused (0)
counter2 = sample in stream

The detector index is now returned as a dictionary from the TOD
class.  The observation index is contained in the observation
dictionary with the id key.  The component and telescope
indices are not yet used.  These changes result in different
realizations of the random timestream values compared to
previous versions of the code, but all statistical and unit
tests still pass.  For runs with multiple observations, this
now ensures reproducibility of the timestreams regardless of
which process group is assigned a given observation.