Implement a new mapmaking template for periodic signals
This adds a new template that regresses components of
detector data that are correlated with a shared data
field whose values are periodic (HWP angle, Azimuth, etc).
The values of the periodic field are binned and a
different set of template amplitudes are solved for
each detector and each observation.  Additional changes:

* Unit test for fake HWPSS removal in satellite and
  ground simulations.

* Unit test helper for healpix map plotting, similar
  to the helper function for WCS plotting.

* Added functions to the Offset template to support
  writing and plotting solved amplitudes.

* Small improvement to noise model fitting.

* Run format_source.sh