Update required language standard to C++14
Our bundled copy of the google test framework now requires C++14.
We could relax this in the future if needed once we reorganize
the compiled extension and ensure that all features are tested
from the python interface.  In that scenario, we could then
completely remove the compiled test runner and gtest.