Add a new, general purpose convolution function
In several places throughout the code we perform Fourier domain convolution
and use a variety of techniques.  This work aims to have a common function
for some of these use cases which carefully handles the symmetric extension
of data in the Fourier domain buffer and is unit tested.  Changes include:

- New toast.fft.convolve function that can take precomputed kernels or
  use a callback function to generate the kernels.  Unit tests that
  verify roundtrip and expected sample phase shift introduced by a
  Butterworth filter.

- Porting of the timeconstant deconvolution operator to this new function.

- Small unrelated fix to toast timing plots

- Unrelated fixes to DetectorData indexing.  In a couple places we were
  triggering data copying by using numpy "advanced indexing".