Add ability to create a PixelDistribution from a sky footprint.
In most workflows, a PixelDistribution is created by passing through
the detector pointing on each process and determining the locally
hit pixels and submaps for a given pixelization scheme.  This work
adds support for creating a distribution from a pre-existing sky
footprint.  Each process will have the same hit submaps in this case.
This is a stepping stone towards building the RHS out of core while
keeping the locally hit pixels in node shared memory.